+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * [#1865](https://github.com/bbatsov/rubocop/issues/1865): New cop `Lint/UnneededDisable` checks for `rubocop:disable` comments that can be removed. ([@jonas054][])
 * `EmptyElse` cop does auto-correction. ([@lumeet][])
 * Show reference links when displaying style guide links. ([@rrosenblum][])
+* Debugger cop now checks for the Capybara debug method save_screenshot. ([@crazydog115][])
 
 ### Changes
 
@@ -1423,3 +1424,4 @@
 [@ojab]: https://github.com/ojab
 [@chastell]: https://github.com/chastell
 [@glasnt]: https://github.com/glasnt
+[@crazydog115]: https://github.com/crazydog115